requirements.txt
langchain~=0.3
langchain-openai~=0.3
langchain-core~=0.3
langchain-community~=0.3
# RAG dependencies
chromadb
langchain-chroma
sentence-transformers
langchain-huggingface
unstructured[md]
nltk
lark
# Open source models dependencies
# langchain-ollama~=0.2
# Dev dependencies
black
python-dotenv
python-docx
